seed more

+++ viking_store/db/migrate/20160802184057_create_products.rb
@@ -4,7 +4,7 @@ class CreateProducts < ActiveRecord::Migration[5.0]
       t.string :title
       t.text :description
       t.float :price
-      t.integer :sku
+      t.bigint :sku
       
       t.integer :category_id
       t.timestamps
